requirements.txt
jinja2
weasyprint
jsonschema
google-genai
python-docx